Relax 'requests' dependency
Let it pick the version from the distribution.
Note the versioned dependency in salt was due to
https://github.com/psf/requests/security/advisories/GHSA-9wx4-h78v-vm56,
so it's not great, but doesn't affect our use case (the verify= option
with possibly non-true value is used when talking to S3, which we don't
use). In any case, let the update be handled by the distribution.